+++ app/[locale]/(default)/page.tsx
@@ -1,18 +1,11 @@
-import { removeEdgesAndNodes } from '@bigcommerce/catalyst-client';
-import { NextIntlClientProvider } from 'next-intl';
-import { getMessages, getTranslations, unstable_setRequestLocale } from 'next-intl/server';
+import { Metadata } from 'next';
+import { notFound } from 'next/navigation';
+import { unstable_setRequestLocale } from 'next-intl/server';
 
-import { getSessionCustomerId } from '~/auth';
-import { client } from '~/client';
-import { graphql } from '~/client/graphql';
-import { revalidate } from '~/client/revalidate-target';
-import { Hero } from '~/components/hero';
-import {
-  ProductCardCarousel,
-  ProductCardCarouselFragment,
-} from '~/components/product-card-carousel';
-import { ContentstackPosts } from '~/contentstack/posts';
+import { getWebpageData } from '~/contentstack/get-webpage-data';
 import { LocaleType } from '~/i18n';
+
+import RenderComponents from '../../../contentstack/components/render-components';
 
 interface Props {
   params: {
@@ -20,72 +13,36 @@ interface Props {
   };
 }
 
-const HomePageQuery = graphql(
-  `
-    query HomePageQuery {
-      site {
-        newestProducts(first: 12) {
-          edges {
-            node {
-              ...ProductCardCarouselFragment
-            }
-          }
-        }
-        featuredProducts(first: 12) {
-          edges {
-            node {
-              ...ProductCardCarouselFragment
-            }
-          }
-        }
-      }
-    }
-  `,
-  [ProductCardCarouselFragment],
-);
+export async function generateMetadata({ params: { locale } }: Props): Promise<Metadata> {
+  const webpage = await getWebpageData({ locale });
+
+  if (!webpage) {
+    notFound();
+  }
+
+  return {
+    title: webpage.seo?.meta_title,
+    description: webpage.seo?.meta_description,
+    keywords: webpage.seo?.keywords,
+  };
+}
 
 export default async function Home({ params: { locale } }: Props) {
-  const customerId = await getSessionCustomerId();
-
   unstable_setRequestLocale(locale);
 
-  const t = await getTranslations({ locale, namespace: 'Home' });
-  const messages = await getMessages({ locale });
+  const webpage = await getWebpageData({ locale });
 
-  const { data } = await client.fetch({
-    document: HomePageQuery,
-    customerId,
-    fetchOptions: customerId ? { cache: 'no-store' } : { next: { revalidate } },
-  });
-
-  const featuredProducts = removeEdgesAndNodes(data.site.featuredProducts);
-  const newestProducts = removeEdgesAndNodes(data.site.newestProducts);
+  if (!webpage) {
+    notFound();
+  }
 
   return (
-    <>
-      <Hero />
-
-      <ContentstackPosts locale={locale} />
-
-      <div className="my-10">
-        <NextIntlClientProvider locale={locale} messages={{ Product: messages.Product ?? {} }}>
-          <ProductCardCarousel
-            products={featuredProducts}
-            showCart={false}
-            showCompare={false}
-            showReviews={false}
-            title={t('Carousel.featuredProducts')}
-          />
-          <ProductCardCarousel
-            products={newestProducts}
-            showCart={false}
-            showCompare={false}
-            showReviews={false}
-            title={t('Carousel.newestProducts')}
-          />
-        </NextIntlClientProvider>
-      </div>
-    </>
+    <RenderComponents
+      contentTypeUid={webpage.system?.content_type_uid}
+      entryUid={webpage.system?.uid}
+      locale={webpage.system?.locale ?? 'en'}
+      pageComponents={webpage.page_components ?? []}
+    />
   );
 }
